Fix intermittent crash when deleting under load
Hack to fix #1885

This doesn't happen often, but is reproducible under simulated message
churn (~50 insert/update/deletes randomly spaced over 10 seconds).

During performBatchUpdates, the collectionViewLayout can call
layoutAttributesForElementsInRect which can reference the just-deleted item's
view. (The view presumably hasn't been deleted yet, but will be by the time
performBatchUpdates is complete).

It's opaque how layoutAttributesForElementsInRect get's it's list of views. I
can only speculate how this works based on debugging, but it seems that the
CollectionFlowLayout keeps an internal list of item frames, which when crashing
still contain items corresponding to the to-be-deleted items.

In any case, it seems like a bug that JSQMVC would ever attempt to get
layoutAttributes for an item which doesn't exist in the datasource.