remove unnecessary keyboard unassignment
Backstory is that in Signal we want to toggle the default keyboard.

To do this we dismiss and then pop the keyboard.

However, the event listeners cause an unsightly bounce animation when
this happens, so we'd temporarily disable the event listeners.

But then there is this side effect of unassigned the keyboard view when
toggling off the event listeners. There seems to be no reason for it.
(famous last words).

The consequence of unassigning the keyboard view was that future
calculations about keyboard dismissal were wrong, preventing you from
being able to dismiss the keyboard after sending a message.